+++ test/modules/test/index.js
@@ -1,6 +1,12 @@
 var gn = require('../../../');
 var logger = gn.log.create('modules/test');
 
+gn.defineOption('-R', 'Option given from mocha', function (val) {
+	var logger = gn.log.create('argv');
+	logger.debug('-R caught from test module:', val);
+});
+
+
 module.exports.readConfig = function () {
 	logger.info('readConfig executed');
 };